pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ifixai"
version = "1.0.0"
description = "ifixai — open-source AI misalignment test suite"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "ifixai Contributors"},
]
keywords = ["ai", "governance", "safety", "compliance", "tests", "misalignment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "click>=8.0",
    "aiohttp>=3.9",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
openrouter = ["openai>=1.0"]
gemini = ["google-generativeai>=0.3"]
anthropic = ["anthropic>=0.18"]
azure = ["openai>=1.0"]
bedrock = ["boto3>=1.28"]
huggingface = ["huggingface-hub>=0.20"]
all = [
    "openai>=1.0",
    "google-generativeai>=0.3",
    "anthropic>=0.18",
    "boto3>=1.28",
    "huggingface-hub>=0.20",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "bandit>=1.7",
    "pre-commit>=3.5",
    "openai>=1.0",
]

[project.scripts]
ifixai = "ifixai.cli.main:main"

[project.urls]
Homepage = "https://github.com/ifixai-ai/diagnostic"
Documentation = "https://github.com/ifixai-ai/diagnostic#readme"
Repository = "https://github.com/ifixai-ai/diagnostic"
Issues = "https://github.com/ifixai-ai/diagnostic/issues"

[tool.pytest.ini_options]
markers = [
    "integration: tests that depend on on-disk fixtures, shipped YAML, or recorded cassettes",
]

[tool.setuptools.packages.find]
include = ["ifixai*"]

[tool.setuptools.package-data]
ifixai = ["py.typed", "fixtures/**/*.yaml", "fixtures/**/*.json", "fixtures/schema.json", "rules/definitions/*.yaml", "mappings/*.yaml", "judge/rubrics/analytic/*.yaml", "judge/references/*.yaml"]
